requirements.txt
absl-py==2.3.1
altair==6.0.0
annotated-types==0.7.0
anyio==4.11.0
argcomplete==3.6.2
asttokens==3.0.1
astunparse==1.6.3
attrs==25.3.0
audioop-lts==0.2.1
azure-ai-documentintelligence==1.0.2
azure-core==1.33.0
azure-identity==1.21.0
babel==2.17.0
backrefs==5.9
beautifulsoup4==4.13.4
bleach==6.3.0
blinker==1.9.0
build==1.3.0
CacheControl==0.14.4
cachetools==6.2.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
cobble==0.1.4
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
coverage==7.12.0
crashtest==0.4.1
cryptography==44.0.2
csscompressor==0.9.5
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
distlib==0.4.0
distro==1.9.0
dotty-dict==1.3.1
dulwich==0.24.10
et_xmlfile==2.0.0
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.0
findpython==0.7.1
Flask==3.1.2
flatbuffers==25.2.10
gast==0.7.0
ghp-import==2.1.0
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.26.0
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.71.0
greenlet==3.3.1
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h5py==3.15.1
halo==0.0.31
hid==1.0.8
hjson==3.1.0
htmlmin2==0.1.13
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.1
importlib_resources==6.5.2
iniconfig==2.3.0
installer==0.7.0
invoke==2.2.1
ipykernel==6.31.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
jsmin==3.0.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
jupytext==1.19.1
keras==3.12.0
keyring==25.7.0
libclang==18.1.1
log-symbols==0.0.14
lxml==5.4.0
magika==0.6.1
mammoth==1.10.0
Markdown==3.8.2
markdown-it-py==4.0.0
markdownify==1.1.0
markitdown==0.1.3
MarkupSafe==3.0.2
matplotlib-inline==0.2.1
mdit-py-plugins==0.5.0
mdurl==0.1.2
mergedeep==1.3.4
mike==2.1.3
milc==1.9.1
mistune==3.2.0
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-git-authors-plugin==0.10.0
mkdocs-git-revision-date-localized-plugin==1.5.1
mkdocs-glightbox==0.5.2
mkdocs-jupyter==0.25.1
mkdocs-macros-plugin==1.5.0
mkdocs-material==9.7.1
mkdocs-material-extensions==1.3.1
mkdocs-minify-plugin==0.8.0
mkdocs-print-site-plugin==2.8
mkdocs-revealjs==0.0.6
mkdocs-rss-plugin==1.17.9
mkdocs-simple-hooks==0.1.5
mkdocs-video==1.5.0
mkdocs_quiz==1.5.4
ml_dtypes==0.5.4
more-itertools==10.8.0
mpmath==1.3.0
msal==1.32.3
msal-extensions==1.3.1
msgpack==1.1.2
mslex==1.3.0
namex==0.1.0
narwhals==2.13.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.2.5
olefile==0.47
onnxruntime==1.20.1
openai==2.6.0
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
paginate==0.5.7
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.6
pathspec==0.12.1
pbs-installer==2025.11.20
pdfminer.six==20250416
pillow==12.1.1
pipx==1.8.0
pkginfo==1.12.1.2
platformdirs==4.3.8
playwright==1.58.0
plotly==6.5.0
pluggy==1.6.0
poetry==2.2.1
poetry-core==2.2.1
polib==1.2.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==5.29.5
psutil==6.1.1
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.12.3
pydantic_core==2.41.4
pydeck==0.9.1
pydub==0.25.1
pyee==13.0.1
pygame==2.6.1
Pygments==2.19.2
PyJWT==2.10.1
pymdown-extensions==10.21
pyparsing==3.2.5
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pyserial==3.5
pytest==9.0.1
pytest-base-url==2.1.0
pytest-cov==7.0.0
pytest-mock==3.15.1
pytest-playwright==0.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-pptx==1.0.2
python-slugify==8.0.4
pytz==2025.2
pyusb==1.3.1
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyyaml_env_tag==1.1
pyzmq==27.1.0
qmk==1.1.8
questionary==2.1.1
RapidFuzz==3.14.3
referencing==0.36.2
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-click==1.9.7
rpds-py==0.26.0
rsa==4.9.1
selectolax==0.4.6
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SpeechRecognition==3.14.2
spinners==0.0.24
stack-data==0.6.3
standard-aifc==3.13.0
standard-chunk==3.13.0
streamlit==1.52.1
super_collections==0.6.2
sympy==1.14.0
taskipy==1.14.1
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow-hub==0.12.0
tensorflowjs==3.18.0
termcolor==3.1.0
termynal==0.13.1
text-unidecode==1.3
tinycss2==1.4.0
toml==0.10.2
tomli==2.4.0
tomlkit==0.13.3
tornado==6.5.3
tqdm==4.67.1
traitlets==5.14.3
trove-classifiers==2025.9.11.17
types-colorama==0.4.15.20250801
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
userpath==1.9.2
verspec==0.1.0
virtualenv==20.35.4
watchdog==6.0.0
wcwidth==0.6.0
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.1
xlrd==2.0.1
XlsxWriter==3.2.3
youtube-transcript-api==1.0.3
zipp==3.23.0
zstandard==0.25.0
